fix: ChatMessageAttachment 엔티티 Integer->Long sizeBytes 변수 수정

+++ src/main/java/com/cotato/itda/domain/chat/entity/ChatMessageAttachment.java
@@ -75,7 +75,7 @@ public class ChatMessageAttachment extends BaseTimeEntity {
 		AttachmentType attachmentType,
 		String objectKey,
 		String mimeType,
-		Integer sizeBytes,
+		Long sizeBytes,
 		AttachmentStatus status,
 		Integer durationMs
 	) {
